Version 1.1.8 released
> Last fix for the selection of the custom editor (when creating a new
node forum) was bad coded ; this release has modified the way to save
the editor choice in the forum node. Everything should be ok now. Sorry
for this.
>Fix a small bug that prevents to update the stopautolinking value when
importing Bb Codes
> The Raw Bb Code xml file has been updated (the bulk file as well).
Reason: the stopautolinking function was not activated

After updating, don't forget to import the Raw Bb Code or to import the
Bb Codes Bulk Xml files (with the override option)